Move sample invoices to api
